added SecurityContextHolder, role based auth and AdminController

+++ src/main/java/com/fridgeMate/fridgemate/service/FridgeItemService.java
@@ -3,9 +3,11 @@ package com.fridgeMate.fridgemate.service;
 import com.fridgeMate.fridgemate.entity.FridgeItem;
 import com.fridgeMate.fridgemate.entity.User;
 import com.fridgeMate.fridgemate.repository.FridgeItemEntryRepository;
+import com.fridgeMate.fridgemate.repository.UserRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.Optional;
@@ -19,11 +21,19 @@ public class FridgeItemService {
     @Autowired
     private UserService userService;
 
+    @Autowired
+    private UserRepository userRepository;
+
+    @Transactional
     public void saveEntry(FridgeItem fridgeItem, String userName) {
-        User user = userService.findByUserName(userName);
-        FridgeItem saved = fridgeItemEntryRepository.save(fridgeItem);
-        user.getFridgeItems().add(saved);
-        userService.saveEntry(user);
+        try {
+            User user = userService.findByUserName(userName);
+            FridgeItem saved = fridgeItemEntryRepository.save(fridgeItem);
+            user.getFridgeItems().add(saved);
+            userService.saveUser(user);
+        } catch (Exception e) {
+            throw new RuntimeException("Error occurred while saving an entry " + e );
+        }
     }
 
     public void saveEntry(FridgeItem fridgeItem) {
@@ -38,11 +48,24 @@ public class FridgeItemService {
         return fridgeItemEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getFridgeItems().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        fridgeItemEntryRepository.deleteById(id);
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removed = user.getFridgeItems().removeIf(x -> x.getId().equals(id));
+
+            if(removed){
+                userService.saveUser(user);
+                fridgeItemEntryRepository.deleteById(id);
+            }
+        } catch (Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while deleting an entry : " + e);
+        }
+        return removed;
     }
 
+    public User findByUserName(String userName) {
+        return userRepository.findByUserName(userName);
+    }
 }
